+++ source/ci_source/providers/BuddyWorks.ts
@@ -26,13 +26,14 @@ export class BuddyWorks implements CISource {
   }
 
   get isPR(): boolean {
-    const mustHave = ["BUDDY_PIPELINE_ID", "BUDDY_EXECUTION_PULL_REQUEST_ID", "BUDDY_REPO_SLUG"]
-    const mustBeInts = ["BUDDY_EXECUTION_PULL_REQUEST_ID"]
+    const mustHave = ["BUDDY_PIPELINE_ID", "BUDDY_EXECUTION_PULL_REQUEST_NO", "BUDDY_REPO_SLUG"]
+    const mustBeInts = ["BUDDY_EXECUTION_PULL_REQUEST_NO"]
+
     return ensureEnvKeysExist(this.env, mustHave) && ensureEnvKeysAreInt(this.env, mustBeInts)
   }
 
   get pullRequestID(): string {
-    return this.env.BUDDY_EXECUTION_PULL_REQUEST_ID
+    return this.env.BUDDY_EXECUTION_PULL_REQUEST_NO
   }
 
   get repoSlug(): string {