+++ source/ci_source/providers/_tests/_buddyWorks.test.ts
@@ -3,7 +3,7 @@ import { getCISourceForEnv } from "../../get_ci_source"
 
 const correctEnv = {
   BUDDY_PIPELINE_ID: "170873",
-  BUDDY_EXECUTION_PULL_REQUEST_ID: "1799",
+  BUDDY_EXECUTION_PULL_REQUEST_NO: "1799",
   BUDDY_REPO_SLUG: "danger/dangerjs",
   BUDDY_EXECUTION_URL:
     "https://app.buddy.works/danger/dangerjs/pipelines/pipeline/170873/execution/5d6d49dbaab2cb6fdf975c71",
@@ -39,7 +39,7 @@ describe(".isPR", () => {
     expect(buddyWorks.isPR).toBeFalsy()
   })
 
-  const envs = ["BUDDY_EXECUTION_PULL_REQUEST_ID", "BUDDY_REPO_SLUG"]
+  const envs = ["BUDDY_EXECUTION_PULL_REQUEST_NO", "BUDDY_REPO_SLUG"]
   envs.forEach((key: string) => {
     const env = Object.assign({}, correctEnv)
     env[key] = null
@@ -52,7 +52,7 @@ describe(".isPR", () => {
 
   it("needs to have a PR number", () => {
     const env = Object.assign({}, correctEnv)
-    delete env.BUDDY_EXECUTION_PULL_REQUEST_ID
+    delete env.BUDDY_EXECUTION_PULL_REQUEST_NO
     const buddyWorks = new BuddyWorks(env)
     expect(buddyWorks.isPR).toBeFalsy()
   })